+++ StactNavigator.js
@@ -1,17 +1,17 @@
-import { AntDesign, Entypo, Ionicons } from "@expo/vector-icons";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { StyleSheet, Text, View } from "react-native";
 import React from "react";
-import BookingScreen from "./screens/BookingScreen";
-import DetailsScreen from "./screens/DetailsScreen";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import HomeScreen from "./screens/HomeScreen";
-import LoginScreen from "./screens/LoginScreen";
-import MapSearchScreen from "./screens/MapSearchScreen";
-import ProfileScreen from "./screens/ProfileScreen";
-import RegisterScreen from "./screens/RegisterScreen";
+import { AntDesign, Entypo, Ionicons } from "@expo/vector-icons";
 import SavedScreen from "./screens/SavedScreen";
-
+import BookingScreen from "./screens/BookingScreen";
+import ProfileScreen from "./screens/ProfileScreen";
+import LoginScreen from "./screens/LoginScreen";
+import RegisterScreen from "./screens/RegisterScreen";
+import DetailsScreen from "./screens/DetailsScreen";
+import CreateListingScreen from "./screens/CreateListingScreen";
+import { NavigationContainer } from "@react-navigation/native";
 const StackNavigator = () => {
   const Tab = createBottomTabNavigator();
   const Stack = createNativeStackNavigator();
@@ -46,26 +46,20 @@ const StackNavigator = () => {
               ),
           }}
         />
-
         <Tab.Screen
           name="Bookings"
           component={BookingScreen}
           options={{
-            tabBarLabel: "Bookings",
+            tabBarLabel: "Map",
             headerShown: false,
             tabBarIcon: ({ focused }) =>
               focused ? (
-                <Ionicons name="notifications" size={24} color="#003580" />
+                <Ionicons name="map" size={24} color="#003580" />
               ) : (
-                <Ionicons
-                  name="notifications-outline"
-                  size={24}
-                  color="black"
-                />
+                <Ionicons name="map-outline" size={24} color="black" />
               ),
           }}
         />
-
         <Tab.Screen
           name="Profile"
           component={ProfileScreen}
@@ -86,26 +80,6 @@ const StackNavigator = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator>
-      <Stack.Screen
-          name="Home"
-          component={HomeScreen}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="Saved"
-          component={SavedScreen}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="Map/Search"
-          component={MapSearchScreen}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-        name="Profile"
-        component={ProfileScreen}
-        options={{ headerShown: false }}
-      />
         <Stack.Screen
           name="Login"
           component={LoginScreen}
@@ -124,12 +98,18 @@ const StackNavigator = () => {
         <Stack.Screen
           name="Details"
           component={DetailsScreen}
-          options={{ headerShown: false }}
+          options={{ headerShown: true }}
+        />
+        <Stack.Screen
+          name="CreateListing"
+          component={CreateListingScreen} // Add the new screen here
+          options={{ headerShown: true, title: "Create Listing" }}
         />
       </Stack.Navigator>
     </NavigationContainer>
   );
 };
+
 export default StackNavigator;
 
 const styles = StyleSheet.create({});
