+++ components/ExploreHeader.js
@@ -1,33 +1,27 @@
+import React from 'react';
+import {
+  View,
+  TextInput,
+  SafeAreaView,
+  StyleSheet,
+  TouchableOpacity,
+} from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
-import { useNavigation } from '@react-navigation/native';
-import { SafeAreaView, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { useNavigation } from '@react-navigation/native'; // Hook for navigation
 
 const ExploreHeader = () => {
-  const navigation = useNavigation();
+ 
+  const navigation = useNavigation(); // Initialize navigation
 
   return (
     <SafeAreaView style={styles.safeArea}>
       <View style={styles.container}>
-        {/* Navigation Buttons Row */}
-        <View style={styles.navigationRow}>
-          <TouchableOpacity onPress={() => navigation.navigate('Home')}>
-            <Ionicons name="home-outline" size={24} color="#003580" />
-            <Text style={styles.navText}>Home</Text>
-          </TouchableOpacity>
-
-          <TouchableOpacity onPress={() => navigation.navigate('Saved')}>
-            <Ionicons name="heart-outline" size={24} color="#003580" />
-            <Text style={styles.navText}>Favorites</Text>
-          </TouchableOpacity>
-
-          <TouchableOpacity onPress={() => navigation.navigate('Map/Search')}>
-            <Ionicons name="map-outline" size={24} color="#003580" />
-            <Text style={styles.navText}>Map</Text>
-          </TouchableOpacity>
-
-          <TouchableOpacity onPress={() => navigation.navigate('Profile')}>
-            <Ionicons name="person-outline" size={24} color="#003580" />
-            <Text style={styles.navText}>Profile</Text>
+        <View style={styles.actionRow}>
+          <TouchableOpacity
+            style={styles.filterBtn}
+            onPress={() => navigation.navigate('CreateListing')} // Navigate to CreateListing
+          >
+            <Ionicons name="create" size={24} />
           </TouchableOpacity>
         </View>
       </View>
@@ -37,9 +31,10 @@ const ExploreHeader = () => {
 
 const styles = StyleSheet.create({
   safeArea: {
-    flex: 0,
+    flex: 1,
     backgroundColor: '#fff',
-    paddingTop: 30,
+    marginBottom: "-145%",
+    marginTop: "6%"
   },
   container: {
     backgroundColor: '#fff',
@@ -47,22 +42,21 @@ const styles = StyleSheet.create({
     shadowColor: '#000',
     shadowOpacity: 0.1,
     shadowRadius: 6,
-    shadowOffset: {
-      width: 1,
-      height: 10,
-    },
+    shadowOffset: { width: 1, height: 10 },
+    alignItems: 'flex-end',
   },
-  navigationRow: {
+  actionRow: {
     flexDirection: 'row',
-    justifyContent: 'space-around',
-    paddingVertical: 10,
-    borderTopWidth: 1,
-    borderColor: '#E2E2E2',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingHorizontal: 24,
+    paddingBottom: 16,
   },
-  navText: {
-    fontSize: 12,
-    marginTop: 4,
-    textAlign: 'center',
+  filterBtn: {
+    padding: 10,
+    borderWidth: 1,
+    borderColor: '#A2A0A2',
+    borderRadius: 24,
   },
 });
