+++ package.json
@@ -23,7 +23,8 @@
     "@types/react": "~18.2.79",
     "expo": "~51.0.28",
     "expo-auth-session": "^5.5.2",
-    "expo-image-picker": "^15.0.7",
+    "expo-crypto": "~13.0.2",
+    "expo-image-picker": "~15.0.7",
     "expo-linking": "^6.3.1",
     "expo-location": "~17.0.1",
     "expo-random": "~14.0.1",
