+++ screens/LoginScreen.js
@@ -10,7 +10,6 @@ const LoginScreen = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const navigation = useNavigation();
-
   const login = () => {
     signInWithEmailAndPassword(auth, email, password)
       .then((userCredential) => {
